size-of-variables

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of char: %d bytes(s)\n", sizeof(a));
-	printf("Size of an int: %d bytes(s)\n", sizeof(b));
-	printf("Size of a long int: %d bytes(s)\n", sizeof(c));
-	printf("Size of a long long int: %d bytes(s)\n", sizeof(d));
+	printf("Size of char: %lu bytes(s)\n", (unsigned long)sizeof(a));
+	printf("Size of an int: %d bytes(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a long int: %d bytes(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a long long int: %d bytes(s)\n", (unsigned long)sizeof(d));
 	printf("Size of a float: %d byes(s)\n", sizeof(e));
 	return (0);
 }
